test/requirements.txt
requests>=2.28.0
psycopg2-binary>=2.9.0
argon2-cffi>=21.0.0
